Math.SternBrocotTree

Description

This module implements the algorithm for branch-wise and generation-wise construction of the n-dimensional Stern-Brocot tree due to Hakan Lennerstad as specified in "The n-dimensional Stern-Brocot tree", Blekige Institute of Technology, Research report No. 2012:04.

Documentation

type Sequence v = (Additive v, Traversable v, Foldable v, Finite v, KnownNat (Size v), Num (v Int), Eq (v Int))

Sequence of n coprime positive integers.

treeToLevel

Arguments

:: Sequence v

=> Int
m

-> Graph (v Int)

Subtree of the n-dimensional Stern-Brocot tree extending down to the mth level (generation). The first level corresponds to the sequence of ones.

branchToSequence

Arguments

:: Sequence v

=> v Int
s

-> Graph (v Int)

Branch of the n-dimensional Stern-Brocot tree leading to the sequence s.